Changed build to use an image for compilation
So that the weavedns binary is built in the environment it will run
in, use the `golang` image to build it, then run it to suck the binary
out. This avoids some awkwardness with cross-compilation (by
subsituting some other awkwardness ..)

This is, in part, a pilot for doing the same with the weave router
build.